finalProj2/requirements.txt
Django==4.0
tensorflow-cpu==2.8.0
numpy==1.21.3
psycopg2==2.9.1
Pillow==8.4.0
